requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
joblib>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
threadpoolctl>=3.1.0
Pillow>=9.5.0  # Version compatible avec lime, alibi et fastai
requests>=2.30.0
psycopg2-binary>=2.9.0
redis>=4.5.0
psutil>=5.9.0
flask>=2.2.0  # Compatible avec Dash
werkzeug>=2.2.0,<2.3.0  # Fixe la plage de version de Werkzeug pour résoudre le conflit
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-login>=0.6.0
flask-wtf>=1.1.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
alembic>=1.10.0
yfinance>=0.2.0
pylint>=2.17.0
pytest>=7.3.0
pytest-cov>=4.1.0
python-telegram-bot>=20.0.0
asyncio>=3.4.0
schedule>=1.2.0
protobuf>=3.20.0,<4.0.0  # Version compatible avec tensorflow 

# AI/ML Dependencies
tensorflow>=2.10.0,<2.16.0
keras>=2.10.0
pytorch-lightning>=2.0.0
torch>=1.13.0,<2.1.0  # Version compatible avec optimum, stable-baselines3 et fastai
torchvision>=0.14.0,<0.16.0
transformers>=4.29.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0
# openai>=0.27.0  # Installé par fix-causalml-install.sh avec version compatible
anthropic>=0.3.0
pydantic>=1.10.0,<2.0.0
dill>=0.3.0
h5py>=3.8.0
sqlalchemy>=2.0.0
optuna-integration[tfkeras]>=3.0.0

# Explainable AI / Ensemble Learning
shap>=0.43.0
lime>=0.2.0
interpret>=0.4.0
interpret-core>=0.4.0
alibi>=0.9.0
dalex>=1.5.0
# econml>=0.14.0  # Commenté car peut dépendre de causalml
# causalml>=0.12.0  # Commenté car géré par le script fix-causalml-install.sh
bayesian-optimization>=1.4.0
mlflow>=2.10.0
wandb>=0.15.0
mlxtend>=0.23.0
eli5>=0.13.0
graphviz>=0.20.0  # Pour la visualisation des arbres
dtreeviz>=2.2.0   # Pour la visualisation des arbres de décision
yellowbrick>=1.5.0  # Pour la visualisation des modèles
plotly_express>=0.4.0  # Pour la visualisation des explications

# Trading Analysis
ta>=0.10.0
python-binance>=1.0.0
ccxt>=3.0.0
statsmodels>=0.14.0
alpha_vantage>=2.3.0
backtrader>=1.9.0
# talib>=0.4.0  # Commenté car difficile à installer, nécessite une installation séparée

# Monitoring & Visualization
plotly>=5.14.0  # Pour stabilité
dash>=2.10.0    # Pour stabilité
prometheus_client>=0.16.0  # Pour la surveillance avec Prometheus
# streamlit>=1.22.0       # Commenté car pas utilisé directement dans l'app mais par Plotly
# bokeh>=3.1.0            # Commenté car pas utilisé directement dans l'app

# Advanced Models
sentence-transformers>=2.2.0
hnswlib>=0.7.0
optimum>=1.9.0
langchain>=0.0.0,<0.1.0
# langchain-openai>=0.0.2  # Installé par fix-causalml-install.sh avec version compatible d'openai

# Nouvelles dépendances pour amélioration de précision
optuna>=3.0.0
hyperopt>=0.2.0
vaderSentiment>=3.3.0
nltk>=3.8.0
scikit-optimize>=0.9.0
gpytorch>=1.11.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
tune-sklearn>=0.5.0
fastai>=2.7.0
# finbert-embedding>=0.1.0  # Commenté car requiert torch==1.1.0 incompatible avec les autres dépendances

# Dépendances incompatibles à installer séparément
# NOTE: Pour utiliser finbert-embedding, créer un environnement virtuel séparé avec torch==1.1.0

# Dépendances principales
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
redis==4.5.5
yfinance==0.2.18
pandas==2.0.2
schedule==1.2.0
tensorflow==2.12.0
scikit-learn==1.2.2
matplotlib==3.7.1
plotly==5.14.1
asyncio==3.4.3
aiohttp==3.8.4
httpx==0.24.1
websockets==11.0.3
bcrypt==4.0.1
pyjwt==2.7.0
email-validator==2.0.0.post2
jinja2==3.1.2
pillow==9.5.0
gunicorn==20.1.0
python-multipart==0.0.6
aiofiles==23.1.0
starlette==0.27.0
passlib==1.7.4
loguru==0.7.0
pytest==7.3.1
coverage==7.2.7
pytest-cov==4.1.0
black==23.3.0
isort==5.12.0
flake8==6.0.0

# Dépendances optionnelles
pymongo==4.3.3
motor==3.1.2
celery==5.3.0
greenlet==2.0.2
ujson==5.7.0
orjson==3.9.0
cryptography==40.0.2
typer==0.9.0
python-jose==3.3.0
